finished a first version to release
