+++ src/outputProcess/CMakeLists.txt
@@ -1,8 +1,6 @@
 add_library(outputProcess 
                         G16Process.cpp
-                        G16Process.hpp
-                        Testing.cpp
-                        Testing.hpp)
+                        G16Process.hpp)
 
 target_include_directories( outputProcess PRIVATE 
         "${CMAKE_CURRENT_SOURCE_DIR}"
